# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CCL API",
        "description": "CCL Backend API Documentation",
        "contact": {
            "name": "YOUNOUSS BORE",
            "email": "bore.younous59@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API pour l'ensemble des services de CCL"
        }
    ],
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lister tous les utilisateurs",
                "operationId": "4ecf0192b4a10534aa238fe447a41d33",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer un utilisateur",
                "operationId": "4a5b6446d347511714140e2910716a5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nomComplet",
                                    "login",
                                    "id_role",
                                    "id_contact"
                                ],
                                "properties": {
                                    "nomComplet": {
                                        "type": "string"
                                    },
                                    "login": {
                                        "type": "string"
                                    },
                                    "idCOD": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "id_role": {
                                        "type": "string"
                                    },
                                    "id_contact": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Récupérer un utilisateur",
                "operationId": "7a38988a7164b5a7a5146fdfb7bd6f99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour un utilisateur",
                "operationId": "e2bcc28b0ea34b94629c2df03f3ab49b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomComplet": {
                                        "type": "string"
                                    },
                                    "login": {
                                        "type": "string"
                                    },
                                    "idCOD": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "id_role": {
                                        "type": "string"
                                    },
                                    "id_contact": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "f9151f08ad5f8ea3797c36e3339af479",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé"
                    }
                }
            }
        },
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion d'un utilisateur",
                "description": "Permet à un utilisateur de se connecter en fournissant son login et son mot de passe",
                "operationId": "d74d2bd1cb4b0bf62b4575e6ba001b49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "younouss23"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "motDePasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Identifiants invalides"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "Modèle d'utilisateur",
                "required": [
                    "id",
                    "nomComplet",
                    "login",
                    "id_role",
                    "id_contact"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nomComplet": {
                        "type": "string",
                        "example": "Younouss Bore"
                    },
                    "login": {
                        "type": "string",
                        "example": "younouss23"
                    },
                    "idCOD": {
                        "type": "string",
                        "example": "COD123456",
                        "nullable": true
                    },
                    "id_role": {
                        "type": "string",
                        "example": "R01"
                    },
                    "id_contact": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    },
                    "contact": {
                        "$ref": "#/components/schemas/Contact"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}